OEL-484: Refactor integration test to existing.
